Make upstream test suite pass with HTTP::Tiny v0.090
HTTP::Tiny v0.090 shipped with a change that meant that calls
to 'www_form_urlencode', which we use to encode query parameters,
no longer sort on keys _and_ values, but only on keys [1]. While
this is arguably a sensible change, it means that the canonical
Amazon Signature Version 4 test suite we include as part of our own,
which relies on specific signatures being generated for specific
requests, no longer passes.

This change makes the test suite use an array reference instead of
a hash reference for the last parameter to 'request', which is then
passed as-is to HTTP::Tiny's 'www_form_urlencode', and allows us to
control the order of the parts that make up the encoded string.

We do not document supporting an array reference to be used in this
case, and this change does not modify this.

Fixes #7.

[1]: https://github.com/Perl-Toolchain-Gang/HTTP-Tiny/commit/9fcfe7cf19ba622fdfdfe0deb9ca99832164f50f